// src/types/flow.ts
export interface TraitServicePort {
    port: number;
    targetPort?: number;
    protocol?: string;
    name?: string;
}

export type TraitServiceType = 'internal' | 'node' | 'public' | 'external';

export interface TraitServiceSpec {
    name?: string;
    type?: TraitServiceType;
    headless?: boolean;
    selector?: Record<string, string>;
    externalName?: string;
    labels?: Record<string, string>;
    ports: TraitServicePort[];
}